[refactor] package structure
