+++ src/main/java/kr/co/common/dto/ErrorDto.java
@@ -1,0 +1,20 @@
+package kr.co.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ErrorDto {
+    private String message;
+
+    public static ErrorDto createErrorDto(Exception e){
+        ErrorDto errorDto = new ErrorDto();
+        errorDto.setMessage(e.getMessage());
+        return errorDto;
+    }
+}
