+++ src/main/java/kr/co/common/exception/CustomException.java
@@ -1,0 +1,16 @@
+package kr.co.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class CustomException extends RuntimeException {
+    private String message;
+    private int statusCode;
+
+}
