--- src/main/java/kr/co/talk/Application.java
@@ -1,14 +1,0 @@
-package kr.co.talk;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-
-
-}
