--- src/main/java/kr/co/talk/exception/CustomException.java
@@ -1,16 +1,0 @@
-package kr.co.talk.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class CustomException extends RuntimeException {
-    private String message;
-    private int statusCode;
-
-}
