--- src/main/java/kr/co/talk/exception/ErrorDto.java
@@ -1,20 +1,0 @@
-package kr.co.talk.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ErrorDto {
-    private String message;
-
-    public static ErrorDto createErrorDto(Exception e){
-        ErrorDto errorDto = new ErrorDto();
-        errorDto.setMessage(e.getMessage());
-        return errorDto;
-    }
-}
