--- src/main/java/kr/co/talk/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package kr.co.talk.repository;
-
-import kr.co.talk.domain.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User,Long> {
-    User findByUserUid(String userUid);
-}
