+++ src/main/java/kr/co/user/service/UserService.java
@@ -1,0 +1,27 @@
+package kr.co.user.service;
+
+import kr.co.user.model.User;
+import kr.co.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import static kr.co.user.dto.SocialKakaoDto.UserInfoDto;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class UserService {
+    private final UserRepository userRepository;
+
+    @Transactional
+    public User createUser(UserInfoDto userInfoDto) {
+        User user = userInfoDto.createUser();
+        return userRepository.save(user);
+    }
+
+    public User findByUserUid(String userUid) {
+        return userRepository.findByUserUid(userUid);
+    }
+
+}
